refactor: errors 패키지 추가
