+++ src/main/java/roomescape/support/exception/BadRequestException.java
@@ -1,5 +1,7 @@
 package roomescape.support.exception;
 
+import roomescape.support.exception.errors.Errors;
+
 public class BadRequestException extends RoomescapeException {
 
     public BadRequestException(Errors errors, Object... args) {
