+++ src/main/java/roomescape/support/exception/ConflictException.java
@@ -1,5 +1,7 @@
 package roomescape.support.exception;
 
+import roomescape.support.exception.errors.Errors;
+
 public class ConflictException extends RoomescapeException {
 
     public ConflictException(Errors errors, Object... args) {
