+++ src/main/java/roomescape/support/exception/GlobalExceptionHandler.java
@@ -7,6 +7,7 @@ import org.springframework.validation.ObjectError;
 import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
+import roomescape.support.exception.errors.RoomescapeErrors;
 
 @RestControllerAdvice
 public class GlobalExceptionHandler {
