+++ src/main/java/roomescape/support/exception/InternalServerException.java
@@ -1,5 +1,7 @@
 package roomescape.support.exception;
 
+import roomescape.support.exception.errors.Errors;
+
 public class InternalServerException extends RoomescapeException {
 
     public InternalServerException(Errors errors, Object... args) {
