+++ src/main/java/roomescape/support/exception/NotFoundException.java
@@ -1,5 +1,7 @@
 package roomescape.support.exception;
 
+import roomescape.support.exception.errors.Errors;
+
 public class NotFoundException extends RoomescapeException {
 
     public NotFoundException(Errors errors, Object... args) {
