+++ src/main/java/roomescape/support/exception/errors/ReservationErrors.java
@@ -1,4 +1,4 @@
-package roomescape.support.exception;
+package roomescape.support.exception.errors;
 
 import lombok.Getter;
